Fix function to_categorical in data_utils.py
The asarray function of numpy will keep ndarray's shape, so the original to_categorical function cannot return the right answer when y is a ndarray with 2-dimension or more. A reshape function is added for changing ndarray to 1-d array. And add a warning when the input array is 3-dimension or more.