fix: ChainCallback sharing default mutable list
In python, the default values of function parameters are instantiated
at function definition time. All calls to that function that use the
default value all point to the same global object. e.g.:

```
def func(x=[]):
   x.append(1)
   print(x)

func() # [1]
func() # [1 , 1]
```

Because of this ChainCallback class potentially all share the same
list of callbacks.

Fix:
The recommended solution is to either set default to None and assign
a new empty object when the variable is None.